requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
dateutils==0.6.12
Deprecated==1.2.13
Django==4.0.8
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-elasticsearch-dsl==7.2.2
django-extensions==3.2.1
django-opensearch-dsl==0.4.1
django-timezone-field==5.0
djangorestframework==3.14.0
drf-yasg==1.21.4
elasticsearch==7.17.6
haversine==2.7.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
numpy==1.23.3
opensearch-dsl==2.0.1
opensearch-py==2.0.0
packaging==21.3
Pillow==9.2.0
postgres==4.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.4
psycopg2-pool==1.1
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.4
redis==4.3.4
reportlab==3.6.11
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
